Reduce duplicated code. Conditional videomaster compile / link
